requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
openai==0.28.0  # Revenir à une version plus ancienne qui supporte 'proxies'
gunicorn==21.2.0
werkzeug==2.3.7